vis-v4: memoize nodes, switch nodeTypes; memoize edges; centralize edge style/geometry; add bridge utils; docs updates; remove stray log; keep tests green
